CIV-15925 changes
