+++ src/main/java/uk/gov/hmcts/reform/civil/callback/CaseEvent.java
@@ -570,7 +570,8 @@ public enum CaseEvent {
     NOTIFY_RESPONSE_TO_QUERY(CAMUNDA),
     NOTIFY_OTHER_PARTY_FOR_RAISED_QUERY(CAMUNDA),
     NOTIFY_OTHER_PARTY_QUERY_HAS_RESPONSE(CAMUNDA),
-    CHANGE_LANGUAGE_PREFERENCE(USER);
+    CHANGE_LANGUAGE_PREFERENCE(USER),
+    NOTIFY_LIP_RESPONDENT1_FOR_RESET_PIN(CAMUNDA);
 
     private final UserType userType;
