+++ src/main/java/uk/gov/hmcts/reform/civil/handler/callback/camunda/notification/NotifyDefendantLipForResetPinHandler.java
@@ -1,0 +1,86 @@
+package uk.gov.hmcts.reform.civil.handler.callback.camunda.notification;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import uk.gov.hmcts.reform.ccd.client.model.AboutToStartOrSubmitCallbackResponse;
+import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
+import uk.gov.hmcts.reform.civil.callback.Callback;
+import uk.gov.hmcts.reform.civil.callback.CallbackHandler;
+import uk.gov.hmcts.reform.civil.callback.CallbackParams;
+import uk.gov.hmcts.reform.civil.callback.CaseEvent;
+import uk.gov.hmcts.reform.civil.config.PinInPostConfiguration;
+import uk.gov.hmcts.reform.civil.model.CaseData;
+import uk.gov.hmcts.reform.civil.notify.NotificationService;
+import uk.gov.hmcts.reform.civil.notify.NotificationsProperties;
+import static uk.gov.hmcts.reform.civil.helpers.DateFormatHelper.DATE;
+import static uk.gov.hmcts.reform.civil.helpers.DateFormatHelper.formatLocalDate;
+import static uk.gov.hmcts.reform.civil.utils.PartyUtils.getPartyNameBasedOnType;
+
+import java.util.List;
+import java.util.Map;
+
+import static uk.gov.hmcts.reform.civil.callback.CallbackType.ABOUT_TO_SUBMIT;
+
+@Service
+@RequiredArgsConstructor
+public class NotifyDefendantLipForResetPinHandler extends CallbackHandler implements NotificationData {
+
+    private static final List<CaseEvent> EVENTS = List.of(CaseEvent.NOTIFY_LIP_RESPONDENT1_FOR_RESET_PIN);
+    public static final String TASK_ID_Respondent1 = "NotifyRespondent1ForResetPin";
+    private static final String REFERENCE_TEMPLATE = "claim-reset-pin-notification-%s";
+    private final NotificationService notificationService;
+    private final PinInPostConfiguration pipInPostConfiguration;
+    private final NotificationsProperties notificationsProperties;
+    private final Map<String, Callback> callBackMap = Map.of(
+        callbackKey(ABOUT_TO_SUBMIT), this::notifyDefendantForResetPin
+    );
+
+    @Override
+    protected Map<String, Callback> callbacks() {
+        return callBackMap;
+    }
+
+    @Override
+    public String camundaActivityId(CallbackParams callbackParams) {
+        return TASK_ID_Respondent1;
+    }
+
+    private CallbackResponse notifyDefendantForResetPin(CallbackParams callbackParams) {
+        CaseData caseData = callbackParams.getCaseData();
+        if (caseData.getRespondent1() != null && caseData.getRespondent1().getPartyEmail() != null) {
+                generateDefendantLipPIPEmail(caseData);
+        }
+        return AboutToStartOrSubmitCallbackResponse.builder()
+            .build();
+    }
+
+    @Override
+    public List<CaseEvent> handledEvents() {
+        return EVENTS;
+    }
+
+    @Override
+    public Map<String, String> addProperties(CaseData caseData) {
+        return Map.of(
+            RESPONDENT_NAME, getPartyNameBasedOnType(caseData.getRespondent1()),
+            CLAIMANT_NAME, getPartyNameBasedOnType(caseData.getApplicant1()),
+            ISSUED_ON, formatLocalDate(caseData.getIssueDate(), DATE),
+            RESPOND_URL, pipInPostConfiguration.getRespondToClaimUrl(),
+            CLAIM_REFERENCE_NUMBER, caseData.getLegacyCaseReference(),
+            CLAIM_16_DIGIT_NUMBER, caseData.getCcdCaseReference().toString(),
+            PIN, caseData.getRespondent1PinToPostLRspec().getAccessCode(),
+            RESPONSE_DEADLINE, formatLocalDate(caseData.getRespondent1ResponseDeadline()
+                                                   .toLocalDate(), DATE),
+            FRONTEND_URL, pipInPostConfiguration.getCuiFrontEndUrl()
+        );
+    }
+
+    private void generateDefendantLipPIPEmail(CaseData caseData) {
+        notificationService.sendMail(
+            caseData.getRespondent1().getPartyEmail(),
+            notificationsProperties.getRespondentDefendantResponseForSpec(),
+            addProperties(caseData),
+            String.format(REFERENCE_TEMPLATE, caseData.getLegacyCaseReference())
+        );
+    }
+}
